handle upcoming asdf deprecations (#126)
* only check for AsdfInFits if asdf<3.0

support for AsdfInFits is moving from asdf
to stdatamodels. asdf 3.0 will not have a
fits_embed submodule or an AsdfInFits object.

https://github.com/asdf-format/asdf/pull/1337

this commit adds a check for the asdf version
before import AsdfInFits and adds a unit test
to exercise this code (that will only run for
asdf < 3.0)

* remove uses of deprecated asdf.resolver

asdf is deprecating resolver in favor of the
new style resource mappings.

https://github.com/asdf-format/asdf/pull/1362

the stnode and validate submodules both accessed the
default resolver to provide the resolver as an argument
to asdf.schema.get_validator. The default for this argument
will allow for schema resolution through internal use
of the default resource mappings (resource manager):

https://github.com/asdf-format/asdf/blob/6b4701277bb25c1d4e8583b88bc2b125be9a36d6/asdf/schema.py#L354

as the rad schemas are registered via a resource mapping
(not a legacy extension) these will be used to resolve
schema urls:

https://github.com/spacetelescope/rad/blob/5b64a95600e0130800a1c13a710c046c2c361a1f/pyproject.toml#L43

* fix style issue

* fix asdf version comparison for development release version format